build(rollup): fix rollup config
