+++ package.json
@@ -42,44 +42,43 @@
   },
   "homepage": "https://github.com/jaebradley/tinder-access-token-generator#readme",
   "dependencies": {
-    "@babel/runtime": "^7.3.1",
+    "@babel/runtime": "^7.4.4",
     "axios": "^0.18.0",
     "puppeteer": "^1.15.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.2.3",
-    "@babel/core": "^7.2.2",
-    "@babel/plugin-transform-async-to-generator": "^7.2.0",
-    "@babel/plugin-transform-runtime": "^7.2.0",
-    "@babel/preset-env": "^7.3.1",
-    "@commitlint/cli": "^7.5.2",
-    "@commitlint/config-angular": "^7.5.0",
-    "@commitlint/prompt": "^7.5.0",
-    "@commitlint/prompt-cli": "^7.5.0",
-    "ajv": "^6.9.1",
+    "@babel/cli": "^7.4.4",
+    "@babel/core": "^7.4.4",
+    "@babel/plugin-transform-runtime": "^7.4.4",
+    "@babel/preset-env": "^7.4.4",
+    "@commitlint/cli": "^7.6.0",
+    "@commitlint/config-angular": "^7.6.0",
+    "@commitlint/prompt": "^7.6.0",
+    "@commitlint/prompt-cli": "^7.6.0",
+    "ajv": "^6.10.0",
     "axios-debug": "0.0.4",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-jest": "^24.7.1",
+    "babel-jest": "^24.8.0",
     "babel-preset-minify": "^0.5.0",
-    "codecov": "^3.1.0",
-    "dotenv": "^7.0.0",
+    "codecov": "^3.4.0",
+    "dotenv": "^8.0.0",
     "es-check": "^5.0.0",
-    "eslint": "^5.13.0",
+    "eslint": "^5.16.0",
     "eslint-config-airbnb-base": "^13.1.0",
-    "eslint-plugin-import": "^2.16.0",
-    "husky": "^2.1.0",
-    "jest": "^24.7.1",
+    "eslint-plugin-import": "^2.17.2",
+    "husky": "^2.2.0",
+    "jest": "^24.8.0",
     "jest-mock-axios": "^3.0.0",
-    "rollup": "^1.1.2",
+    "rollup": "^1.11.3",
     "rollup-plugin-babel": "^4.3.2",
     "rollup-plugin-babel-minify": "^8.0.0",
     "rollup-plugin-babel-runtime-external": "^2.0.0",
-    "rollup-plugin-commonjs": "^9.2.0",
+    "rollup-plugin-commonjs": "^9.3.4",
     "rollup-plugin-filesize": "^6.0.1",
     "rollup-plugin-local-resolve": "^1.0.7",
-    "rollup-plugin-node-resolve": "^4.0.0",
+    "rollup-plugin-node-resolve": "^4.2.3",
     "rollup-plugin-terser": "^4.0.4",
-    "semantic-release": "^15.13.12",
+    "semantic-release": "^15.13.13",
     "travis-deploy-once": "^5.0.11"
   }
 }
